PRE-master/requirements.txt
absl-py==0.11.0
alabaster==0.7.12
albumentations==1.0.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
docutils==0.16
easydict==1.9
einops==0.3.0
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.18
glfw==2.1.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hypothesis==6.13.0
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
imagesize==1.2.0
iniconfig==1.1.1
ipdb==0.13.7
ipykernel==6.0.3
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
linformer==0.2.1
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
matplotlib-inline==0.1.2
mistune==0.8.4
munch==2.5.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.1.0
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.14.0
py==1.10.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyOpenGL @ file:///C:/Users/Administrator/OpenGL/PyOpenGL-3.1.5-cp38-cp38-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/Administrator/OpenGL/PyOpenGL_accelerate-3.1.5-cp38-cp38-win_amd64.whl
pyparsing==2.4.7
PyQt5==5.12
PyQtWebEngine==5.12
pyrsistent==0.18.0
pytest==6.2.4
pytest-instafail==0.3.0
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.3.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
sacred==0.8.2
scikit-image==0.18.2
scipy==1.6.0
Send2Trash==1.7.1
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
Sphinx==4.0.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.4
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tifffile==2021.7.2
toml==0.10.2
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
vit-pytorch==0.20.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
